GIF processing seperated to another thread
